lots of css

+++ app/Controllers/ListController.js
@@ -3,10 +3,10 @@ import STORE from "../store.js"
 
 //TODO Don't forget to render to the screen after every data change.
 function _drawLists() {
-    let template = ""
-    STORE.State.lists.forEach(p => template += p.Template)
-    document.getElementById("listBody").innerHTML = template
-    STORE.saveState() 
+  let template = ""
+  STORE.State.lists.forEach(p => template += p.Template)
+  document.getElementById("listBody").innerHTML = template
+  STORE.saveState()
 }
 
 
@@ -20,7 +20,7 @@ export default class ListController {
   }
 
   //This takes the intail input from the user from the task form and converts it into an object and sends it to service.
-  
+
   addTask(event, id) {
     event.preventDefault()
     let form = event.target
@@ -30,13 +30,29 @@ export default class ListController {
     STORE.saveState()
     _drawLists()
   }
+
   deleteTask(id, task) {
-    confirm("Delete Task?")
-    ListService.deleteTask(id, task)
-    _drawLists()
-    
+    swal({
+      title: "Are you sure?",
+      text: "Once destroyed, its gone forever",
+      icon: "warning",
+      buttons: true,
+      dangerMode: true,
+    })
+      .then((willDelete) => {
+        if (willDelete) {
+          swal("Boom, Task destroyed", {
+            icon: "success",
+          });
+          ListService.deleteTask(id, task)
+          _drawLists()
+        } else {
+          swal("Your task is safe");
+        }
+      });
+
   }
-  
+
 
   createList(event) {
     event.preventDefault();
@@ -52,10 +68,25 @@ export default class ListController {
   }
 
   deleteList(id) {
-    confirm("Delete List")
-    ListService.deletePost(id)
-    _drawLists()
+    swal({
+      title: "Are you sure?",
+      text: "Once destroyed, you will never be able to get it back!",
+      icon: "warning",
+      buttons: true,
+      dangerMode: true,
+    })
+      .then((willDelete) => {
+        if (willDelete) {
+          swal("Boom! Your list has been destroyed", {
+            icon: "success",
+          });
+          ListService.deletePost(id)
+          _drawLists()
+        } else {
+          swal("Your list is safe");
+        }
+      });
   }
 
-  
+
 }
